# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: recovery-auth
    description: Recovery console session operations
  - name: recovery
    description: Recovery endpoint, incident, and evidence operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /recovery/auth/session:
    get:
      operationId: getRecoverySession
      tags: [recovery-auth]
      summary: Check the current recovery console session
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoverySession"
    post:
      operationId: createRecoverySession
      tags: [recovery-auth]
      summary: Unlock the recovery console
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryLogin"
      responses:
        "200":
          description: Authenticated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoverySession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      operationId: deleteRecoverySession
      tags: [recovery-auth]
      summary: Lock the recovery console
      responses:
        "204":
          description: Session cleared
  /recovery/summary:
    get:
      operationId: getRecoverySummary
      tags: [recovery]
      summary: Get recovery fleet status totals
      responses:
        "200":
          description: Fleet summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoverySummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/Action1Unavailable"
  /recovery/readiness:
    get:
      operationId: getAction1Readiness
      tags: [recovery]
      summary: Check Action1 recovery data readiness
      description: Confirms Action1 authentication and recovery read access without returning recovery records
      responses:
        "200":
          description: Safe readiness status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Action1Readiness"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /recovery/devices:
    get:
      operationId: listRecoveryDevices
      tags: [recovery]
      summary: List Action1 endpoints with recovery data
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
            maxLength: 100
        - name: recoveryStatus
          in: query
          required: false
          schema:
            type: string
            maxLength: 50
        - name: agentHealth
          in: query
          required: false
          schema:
            type: string
            maxLength: 80
        - name: integrity
          in: query
          required: false
          schema:
            type: string
            maxLength: 30
        - name: freshness
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, STALE, ALL]
      responses:
        "200":
          description: Matching endpoint records
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryDeviceList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/Action1Unavailable"
  /recovery/devices/{endpointId}:
    get:
      operationId: getRecoveryDevice
      tags: [recovery]
      summary: Get one recovery endpoint with labeled location details
      parameters:
        - name: endpointId
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 160
      responses:
        "200":
          description: Endpoint record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryDevice"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Endpoint not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          $ref: "#/components/responses/Action1Unavailable"
  /recovery/incidents:
    get:
      operationId: listRecoveryIncidents
      tags: [recovery]
      summary: List recovery incidents
      responses:
        "200":
          description: Recovery incidents ordered by most recently updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryIncidentList"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createRecoveryIncident
      tags: [recovery]
      summary: Create an incident and capture current endpoint evidence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryIncidentInput"
      responses:
        "201":
          description: Created incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryIncidentDetail"
        "400":
          description: Invalid incident data or unknown endpoint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          $ref: "#/components/responses/Action1Unavailable"
  /recovery/incidents/{incidentId}:
    get:
      operationId: getRecoveryIncident
      tags: [recovery]
      summary: Get an incident with captured endpoint evidence and audit records
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      responses:
        "200":
          description: Incident detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryIncidentDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Incident not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateRecoveryIncident
      tags: [recovery]
      summary: Update incident ownership, status, case information, or add a note
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryIncidentUpdate"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryIncidentDetail"
        "400":
          description: Invalid incident update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Incident not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /recovery/incidents/{incidentId}/export:
    post:
      operationId: exportRecoveryIncident
      tags: [recovery]
      summary: Generate a controlled incident evidence export
      parameters:
        - $ref: "#/components/parameters/IncidentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecoveryEvidenceExportRequest"
      responses:
        "200":
          description: JSON, CSV, or print-ready HTML evidence export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecoveryEvidenceExport"
            text/csv:
              schema:
                type: string
            text/html:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Incident not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  parameters:
    IncidentId:
      name: incidentId
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: An authorized recovery console session is required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    RateLimited:
      description: Too many failed login attempts
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Action1Unavailable:
      description: Action1 could not be reached or did not accept the API credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    RecoveryLogin:
      type: object
      properties:
        password:
          type: string
          minLength: 1
          maxLength: 1024
      required: [password]
    RecoverySession:
      type: object
      properties:
        authenticated:
          type: boolean
        expiresAt:
          type: ["string", "null"]
          format: date-time
      required: [authenticated, expiresAt]
    RecoveryDevice:
      type: object
      properties:
        endpointId:
          type: string
        computerName:
          type: string
        organizationId:
          type: string
        organizationName:
          type: string
        endpointStatus:
          type: string
        operatingSystem:
          type: string
        lastSeen:
          type: ["string", "null"]
        deviceId:
          type: ["string", "null"]
        agentVersion:
          type: ["string", "null"]
        agentHealth:
          type: ["string", "null"]
        recoveryStatus:
          type: ["string", "null"]
        locationStatus:
          type: ["string", "null"]
        locationIntegrity:
          type: ["string", "null"]
        locationAgeMinutes:
          type: ["string", "null"]
        locationPermission:
          type: ["string", "null"]
        locationUpdated:
          type: ["string", "null"]
        lastAttempt:
          type: ["string", "null"]
        lastSuccess:
          type: ["string", "null"]
        locationSequence:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        accuracy:
          type: ["string", "null"]
        locationQuality:
          type: ["string", "null"]
        locationSource:
          type: ["string", "null"]
        positionSource:
          type: ["string", "null"]
        locationError:
          type: ["string", "null"]
        mapLink:
          type: ["string", "null"]
        mapEmbedUrl:
          type: ["string", "null"]
        locationCoordinates:
          type: ["string", "null"]
        locationSummary:
          type: ["string", "null"]
        isMapSafe:
          type: boolean
      required:
        - endpointId
        - computerName
        - organizationId
        - organizationName
        - endpointStatus
        - operatingSystem
        - lastSeen
        - deviceId
        - agentVersion
        - agentHealth
        - recoveryStatus
        - locationStatus
        - locationIntegrity
        - locationAgeMinutes
        - locationPermission
        - locationUpdated
        - lastAttempt
        - lastSuccess
        - locationSequence
        - latitude
        - longitude
        - accuracy
        - locationQuality
        - locationSource
        - positionSource
        - locationError
        - mapLink
        - mapEmbedUrl
        - locationCoordinates
        - locationSummary
        - isMapSafe
    RecoveryDeviceList:
      type: object
      properties:
        devices:
          type: array
          items:
            $ref: "#/components/schemas/RecoveryDevice"
        refreshedAt:
          type: string
          format: date-time
        source:
          type: string
      required: [devices, refreshedAt, source]
    RecoverySummary:
      type: object
      properties:
        totalDevices:
          type: number
          minimum: 0
        activeLocations:
          type: number
          minimum: 0
        staleLocations:
          type: number
          minimum: 0
        attentionRequired:
          type: number
          minimum: 0
        integrityIssues:
          type: number
          minimum: 0
        refreshedAt:
          type: string
          format: date-time
      required:
        - totalDevices
        - activeLocations
        - staleLocations
        - attentionRequired
        - integrityIssues
        - refreshedAt
    Action1Readiness:
      type: object
      properties:
        status:
          type: string
          enum: [READY, NOT_READY]
        checkedAt:
          type: string
          format: date-time
        message:
          type: string
      required:
        - status
        - checkedAt
        - message
    RecoveryIncident:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        caseNumber:
          type: ["string", "null"]
        owner:
          type: ["string", "null"]
        status:
          type: string
          enum: [OPEN, ESCALATED, RECOVERED, CLOSED]
        reportedAt:
          type: string
          format: date-time
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        endpointCount:
          type: number
          minimum: 0
      required:
        - id
        - title
        - caseNumber
        - owner
        - status
        - reportedAt
        - resolvedAt
        - createdAt
        - updatedAt
        - endpointCount
    RecoveryIncidentInput:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 160
        endpointIds:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 160
        caseNumber:
          type: ["string", "null"]
          maxLength: 100
        owner:
          type: ["string", "null"]
          maxLength: 100
        reportedAt:
          type: ["string", "null"]
          format: date-time
        note:
          type: ["string", "null"]
          maxLength: 4000
      required: [title, endpointIds]
    RecoveryIncidentUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 160
        caseNumber:
          type: ["string", "null"]
          maxLength: 100
        owner:
          type: ["string", "null"]
          maxLength: 100
        status:
          type: string
          enum: [OPEN, ESCALATED, RECOVERED, CLOSED]
        note:
          type: string
          minLength: 1
          maxLength: 4000
    RecoveryIncidentEndpointEvidence:
      type: object
      properties:
        endpointId:
          type: string
        organizationId:
          type: string
        organizationName:
          type: string
        capturedAt:
          type: string
          format: date-time
        sourceRefreshedAt:
          type: string
          format: date-time
        device:
          $ref: "#/components/schemas/RecoveryDevice"
      required:
        - endpointId
        - organizationId
        - organizationName
        - capturedAt
        - sourceRefreshedAt
        - device
    RecoveryIncidentAuditRecord:
      type: object
      properties:
        id:
          type: string
        eventType:
          type: string
        actorLabel:
          type: string
        summary:
          type: string
        endpointId:
          type: ["string", "null"]
        occurredAt:
          type: string
          format: date-time
      required: [id, eventType, actorLabel, summary, endpointId, occurredAt]
    RecoveryIncidentDetail:
      allOf:
        - $ref: "#/components/schemas/RecoveryIncident"
        - type: object
          properties:
            evidence:
              type: array
              items:
                $ref: "#/components/schemas/RecoveryIncidentEndpointEvidence"
            audit:
              type: array
              items:
                $ref: "#/components/schemas/RecoveryIncidentAuditRecord"
          required: [evidence, audit]
    RecoveryIncidentList:
      type: object
      properties:
        incidents:
          type: array
          items:
            $ref: "#/components/schemas/RecoveryIncident"
      required: [incidents]
    RecoveryEvidenceExportRequest:
      type: object
      properties:
        format:
          type: string
          enum: [json, csv, print]
      required: [format]
    RecoveryEvidenceExport:
      type: object
      properties:
        exportId:
          type: string
        schemaVersion:
          type: string
        generatedAt:
          type: string
          format: date-time
        source:
          type: string
        incident:
          $ref: "#/components/schemas/RecoveryIncidentDetail"
        limitations:
          type: array
          items:
            type: string
      required:
        - exportId
        - schemaVersion
        - generatedAt
        - source
        - incident
        - limitations