# lib/api-zod/src/generated/types/recoveryEvidenceExport.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecoveryIncidentDetail } from './recoveryIncidentDetail';

export interface RecoveryEvidenceExport {
  exportId: string;
  schemaVersion: string;
  generatedAt: Date;
  source: string;
  incident: RecoveryIncidentDetail;
  limitations: string[];
}