# lib/api-zod/src/generated/types/recoveryIncidentAuditRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecoveryIncidentAuditRecord {
  id: string;
  eventType: string;
  actorLabel: string;
  summary: string;
  /** @nullable */
  endpointId: string | null;
  occurredAt: Date;
}