# lib/api-zod/src/generated/types/recoveryIncidentDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecoveryIncident } from './recoveryIncident';
import type { RecoveryIncidentAuditRecord } from './recoveryIncidentAuditRecord';
import type { RecoveryIncidentEndpointEvidence } from './recoveryIncidentEndpointEvidence';

export type RecoveryIncidentDetail = RecoveryIncident & {
  evidence: RecoveryIncidentEndpointEvidence[];
  audit: RecoveryIncidentAuditRecord[];
};