# lib/api-zod/src/generated/types/recoveryIncidentEndpointEvidence.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecoveryDevice } from './recoveryDevice';

export interface RecoveryIncidentEndpointEvidence {
  endpointId: string;
  organizationId: string;
  organizationName: string;
  capturedAt: Date;
  sourceRefreshedAt: Date;
  device: RecoveryDevice;
}