# lib/api-zod/src/generated/types/recoveryIncidentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecoveryIncidentInput {
  /**
     * @minLength 3
     * @maxLength 160
     */
  title: string;
  /**
     * @minItems 1
     * @maxItems 50
     * @items.minLength 1
     * @items.maxLength 160
     */
  endpointIds: string[];
  /**
     * @maxLength 100
     * @nullable
     */
  caseNumber?: string | null;
  /**
     * @maxLength 100
     * @nullable
     */
  owner?: string | null;
  /** @nullable */
  reportedAt?: Date | null;
  /**
     * @maxLength 4000
     * @nullable
     */
  note?: string | null;
}